+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Fast travel menu creation and moving options are now relative to the cursor.
 - Perform a risky internal cleanup which should have been done months ago to remove the final remnants of telestep.
 - Fix: build lock now fires on the first tile when moving the cursor
+- Add the vehicle cycler. Shift+v cycles your cursor through vehicles near your character. Intended to make driving around more viable, since you can now find your car again.
 
 # 0.16.33 (2025-12-08)